MT-2 fix base_page
