+++ ui_pages/config.py
@@ -7,10 +7,4 @@ class PagesURL:
     CATALOG_PAGE_URL = 'https://mega.readyscript.ru/'
     DIGMA_TABLET_PAGE_URL = 'https://mega.readyscript.ru/catalog/digma/'
     CHECKOUT_PAGE_URL = 'https://mega.readyscript.ru/checkout/'
-
-
-class ErrorMessages:
-    INCORRECT_ELEMENT_TEXT = 'Incorrect element text'
-    INCORRECT_ELEMENT_TITLE = 'Incorrect element title'
-    INCORRECT_URL_PAGE = 'Incorrect current URL'
-    INCORRECT_NUMBER_OF_ITEMS = 'Incorrect number of items'
+    PRODUCT_PAGE_URL = 'https://mega.readyscript.ru/product/Planshet-Digma-iDx10-8Gb/'
