+++ ui_pages/error_messages.py
@@ -1,0 +1,5 @@
+class ErrorMessages:
+    INCORRECT_ELEMENT_TEXT = 'Incorrect element text'
+    INCORRECT_ELEMENT_TITLE = 'Incorrect element title'
+    INCORRECT_URL_PAGE = 'Incorrect current URL'
+    INCORRECT_NUMBER_OF_ITEMS = 'Incorrect number of items'
